// Creditz-Angular-src/app/model/individual-application.ts
export class IndividualApplication {
  pancard: String;
  loanAmount: number;
  age: number;
  gender: String;
  occupation: String;
  criminalRecord: number;
  bankruptcy: number;
  loanTenure: number;
  personId: number;
  userId: number;
  constructor(
    pancard: String,
    loanAmount: number,
    age: number,
    gender: String,
    occupation: String,
    criminalRecord: number,
    bankruptcy: number,
    loanTenure: number
  ) {}
}